Finalizo el (paso 2: estructura inicial del proyecto) con la estructura correspondiente de carpetas del proyecto. Tambien declare constantes en los archivos de routes e importe en app.js pensando en el crecimiento y mantenibilidad del proyecto, subo al repositorio local para que mi compañero pueda hacer pull y realizar los requisitos pendientes en las rutas, asi mismo efectuara un manejo de errores acerca del run test.
