+++ package.json
@@ -19,6 +19,7 @@
   },
   "homepage": "https://github.com/Karolatf/servidor_backend_parejas#readme",
   "dependencies": {
+    "cors": "^2.8.6",
     "express": "^5.2.1"
   }
 }
