+++ src/routes/tasks.routes.js
@@ -1,0 +1,10 @@
+import express from 'express'
+
+const tasksRouter = express.Router()
+
+
+
+
+
+
+export default tasksRouter;
