+++ src/routes/users.routes.js
@@ -1,0 +1,11 @@
+import express from 'express';
+
+const usersRouter = express.Router()
+
+
+
+
+
+
+
+export default usersRouter;
